Make some PATCH methods idempotent.
If starting a already started PostgreSQL server the API will return an
internal server error (500). This is because the pgapi tries to start
the server regardless of its current state.

This commit makes start, stop and promote requests idempotent.

Instead of just doing what ever was requested we no check the
clusters current state and return 200 if the cluster is already that
state.