Changed _run_command to accept more secure arguments
_run_command accepts a string as an argument.The subprocess.Popen
call within _run_command requires an array which was since
solved with a split().

By moving to arrays the handling now is more robust, hence
helperfunctions around ensuring the syscalls security are obsolete -
in hindsight there was little necessity in the first splace.

This not only allows for more secure commands, it also enables
SR_create to directly inject the conninfo-argument without any need
for quoting.