Re-greened tests and switch to full-stack flask tests
Previously we were calling api-functions directly. this works until request-objects are used.

Expected behaviour is to properly utilize the flask framework as introduced in conftest.py.

Not all tests were translated to the new method, as many do not require this approach. However, new tests should be built this way as there is no functional regression.